+++ sql/install/13-opclass_gist.sql
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Bytes & Brains
+ * Copyright 2019-2020 Bytes & Brains
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,22 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
--- complain if script is sourced in psql, rather than via CREATE EXTENSION
-\echo Use "ALTER EXTENSION h3 UPDATE TO 'unreleased'" to load this file. \quit
-
-CREATE OPERATOR <-> (
-  LEFTARG = h3index,
-  RIGHTARG = h3index,
-  PROCEDURE = h3_distance,
-  COMMUTATOR = <->
-);
-
--- Broken since 1.0.0 on update path
-CREATE OR REPLACE FUNCTION h3_to_geometry(h3index) RETURNS geometry
-  AS $$ SELECT ST_SetSRID(h3_to_geo($1)::geometry, 4326) $$ IMMUTABLE STRICT PARALLEL SAFE LANGUAGE SQL;
-CREATE OR REPLACE FUNCTION h3_to_geography(h3index) RETURNS geography
-  AS $$ SELECT h3_to_geometry($1)::geography $$ IMMUTABLE STRICT PARALLEL SAFE LANGUAGE SQL;
 
 -- ---------- ---------- ---------- ---------- ---------- ---------- ----------
 -- GiST Operator Class (opclass_gist.c)